+++ core/helpers/ArrayHelper.php
@@ -113,7 +113,7 @@ class ArrayHelper extends \yii\helpers\BaseArrayHelper
     /**
      * Search for a Column Value inside a Multidimension array and return the array with the found key.
      *
-     * If several results with the same key value exists, the first result is picked.
+     * Compare to searchColumns() this function return will return the first found result.
      *
      * ```php
      * $array = [
@@ -144,6 +144,8 @@ class ArrayHelper extends \yii\helpers\BaseArrayHelper
     /**
      * Search for columns with the given search value, returns the full array with all valid items.
      *
+     * Compare to searchColumn() this function return will return all found results.
+     *
      * > This function is not casesensitive, which means FOO will match Foo, foo and FOO
      *
      * ```php
@@ -152,7 +154,7 @@ class ArrayHelper extends \yii\helpers\BaseArrayHelper
      *     ['name' => 'nadar', 'userId' => 1],
      * ];
      *
-     * $result = ArrayHelper::searchColumn($array, 'userId', '1');
+     * $result = ArrayHelper::searchColumns($array, 'userId', '1');
      *
      * // output:
      * // array (