+++ core/helpers/Html.php
@@ -4,10 +4,9 @@ namespace luya\helpers;
 
 /**
  * Extend Yii Html Helper.
- * 
+ *
  * @author Basil Suter <basil@nadar.io>
  */
 class Html extends \yii\helpers\BaseHtml
 {
-    
 }